requirements.txt
AppOpener
pywhatkit
python-dotenv
beautifulsoup4
rich
groq
requests
keyboard
pyttsx3
wikipedia
pyautogui
PyQt5
cohere
cryptography
edge_tts
flask
flask-cors
flask-sqlalchemy
google-generativeai
googlesearch-python
imutils
mtranslate
numpy>=1.23
opencv-python
pandas
pillow
psutil
pyaudio
pyjokes
pygame
selenium
spacy
speechrecognition
textblob
transformers
webdriver-manager
geocoder
torch
tensorflow
schedule
moviepy
tf-keras
diffusers
